Create structured data to the type SaleEvent
For example, if you create a markup for your local business (usually this content is on a home page) and your business has commercial discounts or sales, then you can apply the type SaleEvent (the sub-type of the type Event) as embed to the property event. You can check all specifying sub-types of type Event - see below as More specific Types.